backend/requirements_whisper.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# Scientific computing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# Audio processing
librosa==0.10.1
soundfile==0.12.1

# OpenAI
openai==1.3.0

# Whisper dependencies
torch==2.0.1
torchaudio==2.0.2
transformers==4.35.0
openai-whisper==20231117

# Additional utilities
requests==2.31.0
Pillow==10.0.1
